Add dragg contstraint.
Packery has some strange behaviours when the user drags the elements
outside the layout wrapper. So the constraint prevents the user from
dragging the elements outside this wrapper. Furthermore it's more
controllable where the items is dropped. The pointer cursor is just for
indicating the user is in an interaction with the item.