Implement moveBlock method.
This method is require because when a block is move from one layout to
another or from one column to another the inner status of the
layoutmanager is not up to date. So the moveBlock method solves this
problem. It removed the block and generates a new one wich will be
inserted in the target column.