Get next block id based on object size.
A generated id makes conflicts when moving a block in a layout where a
block with the same id already exists. So the block id is generated
based on the oject size. Every time a block is inserted the column is
sure to add a block with a unique id.