Use eventEmitter for triggering events globally.
Requirejs configuration was loaded asynchronously so the configuration
wa not ready for loading dependencies.
So the refactoring makes a web folder with all the dependencies and a
main entry point for the app.
For developping it's no longer nessesary building the whole app.
Requirejs loads the dependencies for each request for getting the
changes made in development.